init: move parser related functions to its own file
