+++ src/devices/build.rs
@@ -6,6 +6,7 @@ fn build_default_init() -> PathBuf {
     let manifest_dir = PathBuf::from(std::env::var_os("CARGO_MANIFEST_DIR").unwrap());
     let libkrun_root = manifest_dir.join("../..");
     let init_src = libkrun_root.join("init/init.c");
+    let parser_src = libkrun_root.join("init/parser.c");
     let fs_src = libkrun_root.join("init/fs.c");
     let timesync_src = libkrun_root.join("init/timesync.c");
 
@@ -19,6 +20,10 @@ fn build_default_init() -> PathBuf {
     println!(
         "cargo:rerun-if-changed={}",
         libkrun_root.join("init/jsmn.h").display()
+    );
+    println!(
+        "cargo:rerun-if-changed={}",
+        libkrun_root.join("init/parser.h").display()
     );
     println!(
         "cargo:rerun-if-changed={}",
@@ -48,6 +53,7 @@ fn build_default_init() -> PathBuf {
         .arg("-o")
         .arg(&init_bin)
         .arg(&init_src)
+        .arg(&parser_src)
         .arg(&fs_src);
 
     if timesync_enabled {
